backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
playwright>=1.41.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx>=0.26.0
aiohttp>=3.9.0
sqlalchemy>=2.0.0
google-generativeai>=0.3.0
ollama>=0.1.7
googlemaps>=4.10.0
overpy>=0.6
rapidfuzz>=3.6.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1
pillow>=10.2.0
numpy>=1.26.0
gTTS>=2.4.0
python-multipart>=0.0.6
alembic>=1.13.0
diffimg>=0.3.0
requests-html>=0.10.0
